1.0.98 Signed-off-by: HY-ZhengWei <HY.ZhengWei@qq.com>

+++ src/org/hy/common/StringHelp.java
@@ -1947,7 +1947,7 @@ public final class StringHelp
                 return true;
             }
             
-            Integer.parseInt(str);
+            Double.parseDouble(str);
             return true;
         } 
         catch(Exception e)
